Pin bundler and rename db-init script
The db-init script is used by circle-ci, but it's also picked up by the
postgres image we rely on for development. The circleci invocation runs
it through sed so renaming it to avoid the .sql filename should keep
postgres happy without annoying circleci. We're also pinning bundler at
this point for the same reason it's pinnned in the app.

Signed-off-by: Ian Allison <iana@pims.math.ca>