Remove staff from recipients and set reply to
The reply_to will be the program coordinator at the event location. It
seems that the staff don't always need to receive these emails, so they
are removed for now. The cc filed is unchanged in this commit, but can
be used in practice to allow arbitrary people (including staff) to
receive the messages based on the application settings. To use the cc,
go into the site settings, and under Email/Location, set
`event_statistics_cc` to a comma separated string of addresses who
should get a CC of the message being sent. If the variable is undefined
for a particular location the site default email address will be used.

Signed-off-by: Ian Allison <iana@pims.math.ca>